fix: change update interval parsing from parseInt to parseFloat
The update interval hours should support fractional values for more granular update scheduling. Using parseFloat instead of parseInt allows decimal values from environment variables.

+++ src/infrastructure/config/TVAddonConfig.js
@@ -101,7 +101,7 @@ export class TVAddonConfig {
         // Configuración del modo automático
         autoM3uUrl: process.env.AUTO_M3U_URL || null,
         enableAutoUpdate: process.env.ENABLE_AUTO_UPDATE === 'true',
-        updateIntervalHours: parseInt(process.env.UPDATE_INTERVAL_HOURS) || 4
+        updateIntervalHours: parseFloat(process.env.UPDATE_INTERVAL_HOURS) || 4
       },
 
       // Configuración de TV en vivo
